+++ app/src/main/java/com/app/music_app/exercises/pages/CompareExercisePage.kt
@@ -22,11 +22,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.app.music_app.music_player.interfaces.AbstractInstrument
-import com.app.music_app.view.colors.AppColor
-import com.app.music_app.view.paino_box.PianoCheckbox
-import com.app.music_app.view.piano_keyboard.PianoKeyboard
-import com.app.music_app.view.play_button.PlayButton
+import com.app.music_app.music_players.interfaces.AbstractInstrument
+import com.app.music_app.components.colors.AppColor
+import com.app.music_app.components.paino_box.PianoCheckbox
+import com.app.music_app.components.piano_keyboard.PianoKeyboard
+import com.app.music_app.components.play_button.PlayButton
 import com.musiclib.notes.Melody
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
